feat: dashboard - add panel - ui updates
